agents_simulation/pyproject.toml
[project]
name = "agents-simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.64",
    "fastmcp>=2.13.0.2",
    "python-dotenv>=1.2.1",
    "strands-agents>=1.14.0",
]
